app-python/requirements.txt
#please put all required Python dependencies for your app here
prometheus-client
requests
python-dotenv
